feat(publish-bundle): Scaffold new project
This should be reviewed as if no automation were used.
To scaffold this project, I copied scripts/repackage.sh and skel/ from
the Endo repository and altered them for fitness in the agoric-sdk
workspace. These scripts save some time but are not evolved to the point
they can be used without modification between uses.